Move license template parsing into submodule
